+++ module/Rest/src/Action/Tag/TagsStatsAction.php
@@ -20,7 +20,7 @@ class TagsStatsAction extends AbstractRestAction
     protected const ROUTE_PATH = '/tags/stats';
     protected const ROUTE_ALLOWED_METHODS = [self::METHOD_GET];
 
-    public function __construct(private TagServiceInterface $tagService)
+    public function __construct(private readonly TagServiceInterface $tagService)
     {
     }